Refactor some dialog
