+++ scripts/custom-ring/custom-ring.js
@@ -1,6 +1,6 @@
 import { getBaseJSON } from "./customRingJsonCfg.js";
 import { MODULE_BASE_PATH, MODULE_ID, effects } from "../const.js";
-import { askToReload } from "../module.js";
+import { askToReload } from "../dialog/askToReloadDialog";
 import { downloadCustomRing } from "./export.js";
 
 export function registerCustomRingSettings() {
